Reduce 64-bit arithmetic in cycle counters
- Convert Timekeeper cycle counters to uint32_t (sufficient for ~20 min at 3.58MHz)
- Keep totalCyclesCount as uint64_t for long-running sessions
- Convert blitter cycles and pixel counters to uint32_t
- Use uint32_t for elapsedCycles in CPU dispatch
- Reduces expensive 64-bit operations on ARM9

+++ src/blitter.cpp
@@ -126,7 +126,7 @@ inline void Blitter::ProcessCycle(uint8_t &colorbus) {
     }
 }
 
-void Blitter::CatchUp(uint64_t cycles) {
+void Blitter::CatchUp(uint32_t cycles) {
     if(cycles == 0) {
         cycles = timekeeper->totalCyclesCount - last_updated_cycle;
     }
